pyproject.toml
[tool.poetry]
name = "unified"
version = "0.1.0"
description = ""
authors = ["raushan"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.7"
djangorestframework = "^3.11.0"
drf-yasg = "^1.17.1"
geopy = "^1.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
